requirements.txt
configobj
click
jinja2
obs-websocket-py
